Minor changing for testing

+++ launch_bkg_sim.py
@@ -32,7 +32,7 @@ def read_par(parfile):
     elif line.startswith("@altitudes"):
       altitudes = list(map(float, line.split(" ")[1:]))
     elif line.startswith("@latitudes"):
-      latitudes = list(map(float, line.split(" ")[1:]))
+      latitudes = list(map(int, line.split(" ")[1:]))
       latitudes = np.linspace(latitudes[0], latitudes[1], latitudes[2])
   print(geom, revanf, mimrecf, source_base, spectra, simtime, latitudes, altitudes)
   return geom, revanf, mimrecf, source_base, spectra, simtime, latitudes, altitudes
